add loadbalancer and refactor trace & log exporter
