+++ exporter/loadbalancingexporter/logs_exporter_test.go
@@ -23,11 +23,44 @@ import (
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.uber.org/zap"
 	"math/rand"
-	"net"
-	"sync/atomic"
 	"testing"
-	"time"
 )
+
+func TestLogExporterStart(t *testing.T) {
+	// prepare
+	config := simpleConfig()
+	params := component.ExporterCreateParams{
+		Logger: zap.NewNop(),
+	}
+	p, err := newTracesExporter(params, config)
+	require.NotNil(t, p)
+	require.NoError(t, err)
+	p.res = &mockResolver{}
+
+	// test
+	res := p.Start(context.Background(), componenttest.NewNopHost())
+	defer p.Shutdown(context.Background())
+
+	// verify
+	assert.Nil(t, res)
+}
+
+func TestLogExporterShutdown(t *testing.T) {
+	// prepare
+	config := simpleConfig()
+	params := component.ExporterCreateParams{
+		Logger: zap.NewNop(),
+	}
+	p, err := newTracesExporter(params, config)
+	require.NotNil(t, p)
+	require.NoError(t, err)
+
+	// test
+	res := p.Shutdown(context.Background())
+
+	// verify
+	assert.Nil(t, res)
+}
 
 func TestConsumeLogs(t *testing.T) {
 	// prepare
@@ -124,126 +157,6 @@ func TestLogBatchWithTwoTraces(t *testing.T) {
 	// verify
 	assert.NoError(t, err)
 	assert.Len(t, sink.Logs, 2)
-}
-
-func TestRollingUpdatesWhenConsumeLogs(t *testing.T) {
-	// this test is based on the discussion in the following issue for this exporter:
-	// https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/1690
-	// prepare
-
-	// simulate rolling updates, the dns resolver should resolve in the following order
-	// ["127.0.0.1"] -> ["127.0.0.1", "127.0.0.2"] -> ["127.0.0.2"]
-	res, err := newDNSResolver(zap.NewNop(), "service-1", "")
-	require.NoError(t, err)
-
-	resolverCh := make(chan struct{}, 1)
-	counter := 0
-	resolve := [][]net.IPAddr{
-		{
-			{IP: net.IPv4(127, 0, 0, 1)},
-		}, {
-			{IP: net.IPv4(127, 0, 0, 1)},
-			{IP: net.IPv4(127, 0, 0, 2)},
-		}, {
-			{IP: net.IPv4(127, 0, 0, 2)},
-		},
-	}
-	res.resolver = &mockDNSResolver{
-		onLookupIPAddr: func(context.Context, string) ([]net.IPAddr, error) {
-			defer func() {
-				counter++
-			}()
-
-			if counter <= 2 {
-				return resolve[counter], nil
-			}
-
-			if counter == 3 {
-				// stop as soon as rolling updates end
-				resolverCh <- struct{}{}
-			}
-
-			return resolve[2], nil
-		},
-	}
-	res.resInterval = 10 * time.Millisecond
-
-	config := &Config{
-		Resolver: ResolverSettings{
-			DNS: &DNSResolver{Hostname: "service-1", Port: ""},
-		},
-	}
-	params := component.ExporterCreateParams{Logger: zap.NewNop()}
-	p, err := newLogsExporter(params, config)
-	require.NotNil(t, p)
-	require.NoError(t, err)
-
-	p.res = res
-
-	var counter1, counter2 int64
-	defaultExporters := map[string]component.LogsExporter{
-		"127.0.0.1": &mockLogsExporter{
-			ConsumeLogsFn: func(ctx context.Context, ld pdata.Logs) error {
-				atomic.AddInt64(&counter1, 1)
-				// simulate an unreachable backend
-				time.Sleep(10 * time.Second)
-				return nil
-			},
-		},
-		"127.0.0.2": &mockLogsExporter{
-			ConsumeLogsFn: func(ctx context.Context, ld pdata.Logs) error {
-				atomic.AddInt64(&counter2, 1)
-				return nil
-			},
-		},
-	}
-
-	// test
-	err = p.Start(context.Background(), componenttest.NewNopHost())
-	require.NoError(t, err)
-	defer p.Shutdown(context.Background())
-	// ensure using default exporters
-	p.updateLock.Lock()
-	p.exporters = defaultExporters
-	p.updateLock.Unlock()
-	p.res.onChange(func(endpoints []string) {
-		p.updateLock.Lock()
-		p.exporters = defaultExporters
-		p.updateLock.Unlock()
-	})
-
-	ctx, cancel := context.WithCancel(context.Background())
-	// keep consuming traces every 2ms
-	consumeCh := make(chan struct{})
-	go func(ctx context.Context) {
-		ticker := time.NewTicker(2 * time.Millisecond)
-		for {
-			select {
-			case <-ctx.Done():
-				consumeCh <- struct{}{}
-				return
-			case <-ticker.C:
-				go p.ConsumeLogs(ctx, randomLogs())
-			}
-		}
-	}(ctx)
-
-	// give limited but enough time to rolling updates. otherwise this test
-	// will still pass due to the 10 secs of sleep that is used to simulate
-	// unreachable backends.
-	go func() {
-		time.Sleep(50 * time.Millisecond)
-		resolverCh <- struct{}{}
-	}()
-
-	<-resolverCh
-	cancel()
-	<-consumeCh
-
-	// verify
-	require.Equal(t, []string{"127.0.0.2"}, res.endpoints)
-	require.Greater(t, atomic.LoadInt64(&counter1), int64(0))
-	require.Greater(t, atomic.LoadInt64(&counter2), int64(0))
 }
 
 func randomLogs() pdata.Logs {
